+++ clinique/src/main/java/com/example/clinique/dto/ConsultationDto.java
@@ -9,13 +9,13 @@ public class ConsultationDto {
     private String motif;
     private String traitement;
     private List<String> fichier;
-    private Integer medecinId;
+    private Long medecinId;
     private String medecinNom;
     private String medecinPrenom;
     private String specialite;
 
     public ConsultationDto(Integer id, LocalDateTime dateConsultation, String motif, String traitement,
-                          List<String> fichier, Integer medecinId, String medecinNom, String medecinPrenom, String specialite) {
+                          List<String> fichier, Long medecinId, String medecinNom, String medecinPrenom, String specialite) {
         this.id = id;
         this.dateConsultation = dateConsultation;
         this.motif = motif;
@@ -33,7 +33,7 @@ public class ConsultationDto {
     public String getMotif() { return motif; }
     public String getTraitement() { return traitement; }
     public List<String> getFichier() { return fichier; }
-    public Integer getMedecinId() { return medecinId; }
+    public Long getMedecinId() { return medecinId; }
     public String getMedecinNom() { return medecinNom; }
     public String getMedecinPrenom() { return medecinPrenom; }
     public String getSpecialite() { return specialite; }
